utilsにloss.pyを作成
